+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## HEAD
 
+* Implement max width for modal via data attribute
+
 * Fix vertical resizing for large pictures
 * Fix margins around modal in resize plugin